pyproject.toml
[project]
name = "mantra-db-exporter"
version = "0.1.0"
description = "Firebase Firestore database structure exporter"
requires-python = ">=3.8"
dependencies = [
    "firebase-admin>=6.0.0",
    "google-cloud-firestore>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
